+++ features/merchant/review_inbox/src/main/java/com/tokopedia/review/feature/inbox/buyerreview/domain/model/inboxdetail/ImageAttachmentDomain.kt
@@ -4,6 +4,6 @@ package com.tokopedia.review.feature.inbox.buyerreview.domain.model.inboxdetail
  * @author by nisie on 8/23/17.
  */
 class ImageAttachmentDomain constructor(
-    val attachmentId: Int, val description: String,
+    val attachmentId: Long, val description: String,
     val uriThumbnail: String, val uriLarge: String
 )
