+++ features/merchant/review_inbox/src/main/java/com/tokopedia/review/feature/inbox/buyerreview/view/uimodel/inboxdetail/ImageAttachmentUiModel.kt
@@ -4,7 +4,7 @@ package com.tokopedia.review.feature.inbox.buyerreview.view.uimodel.inboxdetail
  * @author by nisie on 8/24/17.
  */
 data class ImageAttachmentUiModel(
-    var attachmentId: Int,
+    var attachmentId: Long,
     var description: String,
     var uriThumbnail: String,
     var uriLarge: String
